requirements.txt
asgiref==3.5.2
cffi==1.15.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
frozenlist==1.3.1
idna==2.8
importlib-metadata==4.2.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.1
mccabe==0.6.1
mixer==7.1.2
more-itertools==8.14.0
multidict==6.0.2
mypy==0.942
mypy-extensions==0.4.3
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
pep8-naming==0.12.1
Pillow==9.0.0
pipreqs==0.4.11
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==13.14
python-telegram-bot-calendar==1.0.5
python3-openid==3.2.0
pytils==0.4.1
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
testfixtures==6.18.5
toml==0.10.2
tomli==2.0.1
tornado==6.1
typed-ast==1.5.3
typing_extensions==4.2.0
tzdata==2022.4
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.5
webcolors==1.11.1
yarg==0.1.9
yarl==1.8.1
zipp==3.8.0
